Fix auto-discovery of OIDC details to be too eager
For custom providers it was not possible to overwrite
the discovered default settings, because the discovery would be
re-triggered every time that any update was performed,
thereby overwriting the custom settings just entered.

The only way out so far was to completely remove the metadata_url
from the provider.

For custom providers the discovery now only takes place on the
form where the metadata_url is being entered, not on other forms.
For Google and Microsoft discovery remains active on all forms.
Due to further implementation details this means that on creation
they will first fetch their metadata when saving the client details
form, while during edit they will also refetch on the name form
(where Entra allows changing the tenant, which influences the URLs).